Add Cocos3D Xcode templates for Cocos2D v3 and support for arm64.
Add Cocos3D Xcode templates for Cocos2D v3. Refactor all Cocos3D Xcode templates
and restructure template folders. Modify install_cocos3d.sh script to conform to
new template and folder organization. Support arm64 when using Cocos2D v3.
Fix cast issues related to 64-bit compilation in demos.

+++ Projects/CC3Performance/CC3Performance/Classes/CC3PerformanceScene.m
@@ -357,7 +357,7 @@
 
 -(void) prevNodeType {
 	_templateIndex--;
-	if (_templateIndex < 0) _templateIndex = _availableTemplateNodes.count - 1;
+	if (_templateIndex < 0) _templateIndex = (GLint)_availableTemplateNodes.count - 1;
 	self.templateNode = [_availableTemplateNodes objectAtIndex: _templateIndex];
 }
